Collect bundle stats from invalid alleles as well

+++ dsa/src/pos_stats.h
@@ -2,11 +2,11 @@
 #define POS_STATS_H_
 
 typedef struct pos_final_stats_t {
-    double asxs = 0.0;
-    double nm = 0.0;
-    double proper_pairs = 0.0;
+    int asxs = 0;
+    float nm = 0.0f;
+    int proper_pairs = 0;
     // Duplex-only
-    double clip = 0.0;
+    int clip = 0;
 } pos_final_stats_t;
 
 #endif
